backend/requirements.txt
Django
psycopg
python-dotenv
django-cors-headers
djangorestframework